requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
ccxt>=4.0.0
ta>=0.10.2
pyyaml>=6.0
aiofiles>=23.0.0
aiosmtplib>=2.0.0

# Data analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Web and HTTP
aiohttp>=3.8.0
requests>=2.28.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.0.290

# Jupyter for research
jupyter>=1.0.0
matplotlib>=3.7.0
plotly>=5.17.0
